+++ read_survey_xls.py
@@ -501,8 +501,8 @@ if args.ref_fnam is not None:
     if not 'EastingI' in df.columns or not 'NorthingI' in df.columns:
         raise ValueError('Error, identified coordinates not found >>> {}'.format(args.ref_fnam))
     loc_ref = df['Location'].astype(str).str.strip().str.lower().values
-    number_ref = df['BunchNumber'].values
-    plot_ref = df['PlotPaddy'].values
+    number_ref = df['BunchNumber'].astype(int).values
+    plot_ref = df['PlotPaddy'].astype(int).values
     x_ref = df['EastingI'].astype(float).values
     y_ref = df['NorthingI'].astype(float).values
     plant_ref = pd.to_datetime(df['PlantDate']).dt.to_pydatetime()
